Move bottlenecks detection near to security and fix tests
