+++ lib/inquisition/outputter/doc/tpl/bottlenecks_detection.rb
@@ -1,7 +1,3 @@
-require_relative 'security/bottlenecks_detection/active_record_doctor'
-require_relative 'security/bottlenecks_detection/lol_dba'
-require_relative 'security/bottlenecks_detection/fasterer'
-
 module Inquisition
   module Outputter
     class Doc
@@ -14,22 +10,20 @@ module Inquisition
           def active_record_doctor
             @active_record_doctor ||= begin
               Template.new('bottlenecks_detection/active_record_doctor').render(
-                Security::BottlenecksDetection::ActiveRecordDoctor.call(@issues)
+                BottlenecksDetection::ActiveRecordDoctor.call(@issues)
               )
             end
           end
 
           def lol_dba
             @lol_dba ||= begin
-              Template.new('bottlenecks_detection/lol_dba').render(Security::BottlenecksDetection::LolDba.call(@issues))
+              Template.new('bottlenecks_detection/lol_dba').render(BottlenecksDetection::LolDba.call(@issues))
             end
           end
 
           def fasterer
             @fasterer ||= begin
-              Template.new('bottlenecks_detection/fasterer').render(
-                Security::BottlenecksDetection::Fasterer.call(@issues)
-              )
+              Template.new('bottlenecks_detection/fasterer').render(BottlenecksDetection::Fasterer.call(@issues))
             end
           end
         end
@@ -37,3 +31,7 @@ module Inquisition
     end
   end
 end
+
+require_relative 'bottlenecks_detection/active_record_doctor'
+require_relative 'bottlenecks_detection/lol_dba'
+require_relative 'bottlenecks_detection/fasterer'
