+++ lib/inquisition/outputter/doc/tpl/security.rb
@@ -18,19 +18,12 @@ module Inquisition
               Template.new('security/patch_level').render(PatchLevel.call(@issues))
             end
           end
-
-          def bottlenecks_detection
-            @bottlenecks_detection ||= begin
-              Template.new('security/bottlenecks_detection').render(TPL::BottlenecksDetection.new(@issues))
-            end
-          end
         end
       end
     end
   end
 end
 
-require_relative 'bottlenecks_detection'
 require_relative 'security/collector'
 require_relative 'security/brakeman'
 require_relative 'security/patch_level'
