+++ spec/inquisition/bundler/audit/vulnerability_spec.rb
@@ -98,7 +98,6 @@ RSpec.describe Inquisition::Bundler::Audit do
     describe '#to_h' do
       subject(:insecure_source) { described_class.new(source) }
 
-      let(:bundler) { instance_double(Bundler::LazySpecification, source: source) }
       let(:source) { 'http://rubygems.org/' }
       let(:options) do
         {
