+++ spec/inquisition/outputter/doc/tpl/bottlenecks_detection/active_record_doctor_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Inquisition::Outputter::Doc::TPL::Security::BottlenecksDetection::ActiveRecordDoctor do
+RSpec.describe Inquisition::Outputter::Doc::TPL::BottlenecksDetection::ActiveRecordDoctor do
   include_examples 'produceable' do
     subject(:tpl) { described_class.new([]) }
   end
