+++ views/doc/bottlenecks_detection.rhtml
