+++ views/doc/bottlenecks_detection/fasterer.rhtml
@@ -1,9 +1,9 @@
 <li>
   <%= "there are #{collection.count} recommendations to make the application faster (the full report can be found here). Where:" %>
   <ul type="circle">
-    <% collection.group do |data, issues_count| %>
+    <% collection.group do |offense, count| %>
       <li>
-        <%= "#{issues_count} of them related to the <a href=#{data[:link]}>#{data[:title]}</a>;" %>
+        <%= "#{count} of them related to the <a href=#{offense[:link]}>#{offense[:title]}</a>;" %>
       </li>
     <% end %>
   </ul>
