+++ views/doc/layout.rhtml
@@ -106,9 +106,11 @@
         <span class="subtitle"><a id="security_issues_detection">Security issues detection</a></span>
 
         <%= security %>
-        <br style="page-break-before: always">
       </div>
 
+      <%= bottlenecks_detection %>
+
+      <br style="page-break-before: always">
       <div class="">
         <div class="mt-5">
           <span class="subtitle"><a id="structural_similarities_analysis_and_code_smell_detection">Structural similarities analysis and code smell detection</a></span>
