+++ views/doc/security.rhtml
@@ -10,5 +10,3 @@
     <li>there is no access control gem (e.g.  pundit, action_policy)</li>
   </ul>
 </div>
-
-<%= bottlenecks_detection %>
